requirements.txt
foxglove-sdk
pyserial
numpy
depthai-sdk
opencv-python
mamba-ssm

